pyproject.toml
[project]
name = "beam-college-2025-hackathon"
version = "0.1.0"
description = "Beam & Dataflow AI Agents"
requires-python = ">=3.12"
dependencies = [
    "apache-beam[gcp,interactive]==2.65.0",
    "google-cloud-bigquery==3.32.0",
    "google-genai==1.15.0",
    "streamlit==1.45.1",
    "temporalio==1.10.0",
]
